favList is a state in App.js and initialized in Login
