+++ frontend/src/components/App.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import {
   Switch,
   Route
@@ -11,22 +11,23 @@ import SavedSchedulesPage from './SavedSchedulesPage';
 import { AuthProvider } from "../contexts/AuthContext";
 
 function App() {
+  const [favList, setFavList] = useState([]);
   return (
     <div>
       <Header/>
         <AuthProvider>
           <Switch>
             <Route path="/search">
-              <ClassSearchPage />
+              <ClassSearchPage favList={favList} setFavList={setFavList} />
             </Route>
             <Route path="/generate">
-              <GenerateSchedulesPage />
+              <GenerateSchedulesPage favList={favList} setFavList={setFavList} />
             </Route>
             <Route path="/saved">
               <SavedSchedulesPage />
             </Route>
             <Route path="/">
-              <LoginPage />
+              <LoginPage setFavList={setFavList}/>
             </Route>
           </Switch>
         </AuthProvider>
