+++ frontend/src/components/Login.js
@@ -1,6 +1,7 @@
 import React, { useState} from 'react'
 import { useAuth } from "../contexts/AuthContext";
 import { useHistory } from "react-router-dom";
+import { auth, db } from "../firebase";
 
 function Login(props) {
     const [email, setEmail] = useState('')
@@ -22,7 +23,24 @@ function Login(props) {
             props.setErrorDisplay(true);
             props.setErrorContent(error.message);
         }
-        setLoading(false)
+        setLoading(false);
+
+        // get favorites list (copied from AuthContext)
+        var docRef = db.collection("users").doc(auth.currentUser.uid);
+        var list = [];
+        docRef.get().then(function(doc) {
+            if (doc.exists) {
+                console.log("favorites retrieved");
+                list = doc.data().favorList;
+                props.setFavList(list);
+            } else {
+                console.error("favorites document dne");
+                props.setFavList([]);
+            }
+        }).catch(() => {
+            console.error("favorites not found");
+            props.setFavList([]);
+        });
     }
 
     return (
